requirements.txt
streamlit>=1.36
segno>=1.6
pillow>=10.3
numpy>=1.26
